ldap: cleanups
move dns resolution to simple only

remove the fallback as we use simple serverset, sysadmin can manually
modify the settings later.

make sure we use the connection setting we used within the profile.

fix sasl user resolution so it find the user and not other objects.

do not change object state in determine*

and probably more.